requirements.txt
Flask
Flask-Session
requests
firebase_admin
flask-wtf
flask-recaptcha
phonenumbers
email_validator
pyrebase4
jinja2==3.0.0